% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paket_alle_Funktionen_Basis_v01_2-0Okt2018.R
\docType{data}
\name{keratosis}
\alias{keratosis}
\title{Simulated Data Set for a Study of Actinic Keratosis Treatments}
\format{A data frame.}
\usage{
data(keratosis)
}
\description{
A data set simulated under the planning assumptions for a study comparing four radiation regimens for a photodynamic 
treatment of actinic keratosis. Each patient receives each treatment in a different skin patch and each patch contains four lesions.
Variables are patient identifier (id), treatment (trt), lesion identifier within a patient (lesion), the binary outcome clearance success (1=success, 0=no success)
reported for each lesion and the metric outcome pain (larger values indicating more pain) reported for each skin patch. The aim of the study is to compare
the treatments B, C and D to the reference treatment A in terms of both outcomes.
}
\examples{
data(keratosis)
head(keratosis)
}
\keyword{datasets}
